errores encontrados y corregidos

+++ src/excepciones/negocio/SaldoInsuficienteException.java
@@ -6,14 +6,14 @@ package excepciones.negocio;
 import excepciones.base.BancoException;
 
 /**
- * @author Pedro
+ * @author Pablo
  *
  */
 public class SaldoInsuficienteException extends BancoException {
 
 	private static final long serialVersionUID = 1L; // <- Añade esto
 
-	public SaldoInsuficienteException(double saldoActual, double solicitado) {
+	public SaldoInsuficienteException(float saldoActual, float solicitado) {
 		super("Saldo insuficiente. Intentaste retirar " + solicitado + " € pero solo hay " + saldoActual + " €.",
 				"ERR_NEG_001");
 	}
